Enhanced template
- removed redundant files
- fixed and added comments
- refactored code
- fixed src structure
- added better readme
- updated packages
