+++ .eslintrc.json
@@ -1,61 +1,85 @@
 {
   "env": {
-    "es6": true,
-    "browser": true
+    "browser": true,
+    "es2021": true
   },
   "extends": [
+    "plugin:react/recommended",
+    "airbnb",
     "airbnb-typescript"
   ],
   "parser": "@typescript-eslint/parser",
   "parserOptions": {
-    "ecmaVersion": 6,
-    "sourceType": "module",
-    "project": "./tsconfig.json",
     "ecmaFeatures": {
       "jsx": true
-    }
+    },
+    "ecmaVersion": 12,
+    "project": "./tsconfig.json",
+    "sourceType": "module"
   },
   "plugins": [
+    "react",
     "@typescript-eslint"
   ],
   "rules": {
-    "linebreak-style": 0,
-    "object-curly-spacing": [ "error", "always" ],
-    "@typescript-eslint/quotes": [ "error", "double" ],
-    "react/jsx-curly-spacing": "off",
-    "@typescript-eslint/comma-dangle": [ "error", "never" ],
-    "object-shorthand": [ "error", "never" ],
-    "@typescript-eslint/semi": [ "error", "always" ],
-    "no-trailing-spaces": [
-      "error", {
-        "skipBlankLines": true,
-        "ignoreComments": true
+    "@typescript-eslint/comma-dangle": [
+      "error",
+      "never"
+    ],
+    "@typescript-eslint/indent": [
+      "error",
+      2
+    ],
+    "@typescript-eslint/lines-between-class-members": [
+      "error",
+      "always",
+      {
+        "exceptAfterSingleLine": true
       }
     ],
-    "eol-last": [ "error", "never" ],
-    "padded-blocks": [
-      "error", {
-        "classes": "always",
-        "blocks": "never",
-        "switches": "never"
+    "@typescript-eslint/naming-convention": [
+      "error",
+      {
+        "selector": "variable",
+        "format": [
+          "camelCase",
+          "UPPER_CASE",
+          "PascalCase"
+        ]
       }
     ],
-    "import/order": "off",
-    "react/jsx-boolean-value": "off",
-    "object-curly-newline": [ "error", {
-      "ObjectExpression": { "multiline": true, "minProperties": 3, "consistent": true },
-      "ImportDeclaration": "never",
-      "ExportDeclaration": "never"
-    } ],
-    "react/jsx-tag-spacing": [ "error", { "beforeSelfClosing": "never" } ],
-    "import/no-extraneous-dependencies": "off",
-    "import/no-duplicates": "off",
-    "no-duplicate-imports": "error",
-    "@typescript-eslint/lines-between-class-members": [ "error", "always", { "exceptAfterSingleLine": true } ],
-    "react/destructuring-assignment": [ "error", "always" ],
-    "no-extra-semi": "error",
-    "implicit-arrow-linebreak": "off",
+    "@typescript-eslint/no-unused-expressions": [
+      "error",
+      {
+        "allowShortCircuit": true,
+        "allowTernary": true
+      }
+    ],
+    "@typescript-eslint/quotes": [
+      "error",
+      "double"
+    ],
+    "@typescript-eslint/semi": [
+      "error",
+      "always"
+    ],
+    "array-bracket-spacing": "off",
     "arrow-body-style": "off",
+    "arrow-parens": [
+      "error",
+      "as-needed"
+    ],
+    "consistent-return": "off",
+    "eol-last": [
+      "error",
+      "never"
+    ],
+    "implicit-arrow-linebreak": "off",
+    "import/no-duplicates": "off",
+    "import/no-extraneous-dependencies": "off",
+    "import/order": "off",
+    "indent": "off",
+    "linebreak-style": 0,
     "max-len": [
       "error",
       {
@@ -69,30 +93,98 @@
         "ignoreRegExpLiterals": true
       }
     ],
-    "no-tabs": [ "error", { "allowIndentationTabs": true } ],
-    "indent": "off",
-    "@typescript-eslint/indent": [ "error", 2 ],
+    "no-duplicate-imports": "error",
+    "no-extra-semi": "error",
+    "no-multiple-empty-lines": [
+      "error",
+      {
+        "max": 1,
+        "maxBOF": 0,
+        "maxEOF": 0
+      }
+    ],
+    "no-param-reassign": [
+      "error",
+      {
+        "props": false
+      }
+    ],
+    "no-plusplus": [
+      "error",
+      {
+        "allowForLoopAfterthoughts": true
+      }
+    ],
+    "no-tabs": [
+      "error",
+      {
+        "allowIndentationTabs": true
+      }
+    ],
+    "no-trailing-spaces": [
+      "error",
+      {
+        "skipBlankLines": true,
+        "ignoreComments": true
+      }
+    ],
+    "object-curly-newline": [
+      "error",
+      {
+        "ObjectExpression": {
+          "multiline": true,
+          "minProperties": 3,
+          "consistent": true
+        },
+        "ImportDeclaration": "never",
+        "ExportDeclaration": "never"
+      }
+    ],
+    "object-curly-spacing": [
+      "error",
+      "always"
+    ],
+    "object-shorthand": [
+      "error",
+      "never"
+    ],
     "operator-linebreak": "off",
-    "react/jsx-wrap-multilines": "off",
+    "padded-blocks": [
+      "error",
+      {
+        "classes": "always",
+        "blocks": "never",
+        "switches": "never"
+      }
+    ],
+    "react/destructuring-assignment": [
+      "error",
+      "always"
+    ],
+    "react/jsx-boolean-value": "off",
     "react/jsx-curly-newline": "off",
-    "no-plusplus": [ "error", { "allowForLoopAfterthoughts": true } ],
-    "array-bracket-spacing": "off",
-    "@typescript-eslint/no-unused-expressions": [ "error", { "allowShortCircuit": true, "allowTernary": true } ],
+    "react/jsx-curly-spacing": "off",
     "react/jsx-props-no-spreading": "off",
-    "no-multiple-empty-lines": [ "error", { "max": 1, "maxBOF": 0, "maxEOF": 0 } ],
+    "react/jsx-tag-spacing": [
+      "error",
+      {
+        "beforeSelfClosing": "never"
+      }
+    ],
+    "react/jsx-wrap-multilines": "off",
     "react/prop-types": "off",
-    "arrow-parens": ["error", "as-needed"],
-    "consistent-return": "off",
-    "no-param-reassign": ["error", { "props": false }],
-    "@typescript-eslint/naming-convention": [ "error", { "selector": "variable", "format": [ "camelCase", "UPPER_CASE", "PascalCase" ] } ],
-    "require-jsdoc": ["error", {
-      "require": {
+    "react/require-default-props": "off",
+    "require-jsdoc": [
+      "error",
+      {
+        "require": {
           "FunctionDeclaration": true,
           "MethodDefinition": true,
           "ClassDeclaration": true,
           "ArrowFunctionExpression": true,
           "FunctionExpression": false
+        }
       }
-    }]
+    ]
   }
 }
