--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "api-spec-template"]
-	path = api-spec-template
-	url = https://github.com/Metatavu/api-spec-template.git
