+++ README.md
@@ -1,12 +1,46 @@
-# Getting Started with Create React App
+# Getting started with React Typescript template
 
-This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
+### `Add OpenApi specification file (swagger.yaml)`
+
+File can be added as a Git submodule, which helps the development with different spec versions:
+
+```sh
+  git submodule add git@github.com:example-api-spec.git
+```
+
+### `Update api spec generator command`
+
+Change the command inside *package.json* to match the added API specification file location
+
+```json
+  ...
+  "scripts": {
+    ...
+    "build-client": "... -i example-api-spec/swagger.yaml ..."
+  }
+```
+
+### `Add environment variables`
+
+To add necessary environment variables, create .env file to project root folder and write variables inside it with following syntax:
+
+```
+REACT_APP_EXAMPLE_ENVIRONMENT_VARIABLE=variable-value
+REACT_APP_EXAMPLE_API_URL=https://example-api.example.com
+...
+```
+
+### `Update public folder contents to match the project`
+
+* Change title and description in *index.html*
+* Update *favicon.ico*, *logo192.png* and *logo512.png*
+* Update names and icon references in *manifest.json*
 
 ## Available Scripts
 
 In the project directory, you can run:
 
-### `yarn start`
+### `npm start`
 
 Runs the app in the development mode.\
 Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
@@ -14,12 +48,7 @@ Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
 The page will reload if you make edits.\
 You will also see any lint errors in the console.
 
-### `yarn test`
-
-Launches the test runner in the interactive watch mode.\
-See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
-
-### `yarn build`
+### `npm run build`
 
 Builds the app for production to the `build` folder.\
 It correctly bundles React in production mode and optimizes the build for the best performance.
@@ -29,18 +58,11 @@ Your app is ready to be deployed!
 
 See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
 
-### `yarn eject`
+### `npm run lint`
 
-**Note: this is a one-way operation. Once you `eject`, you can’t go back!**
+Fixes all linting problems reported by ESLint in the project.
+**Always confirm the changes afterwards!**
 
-If you aren’t satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
+### `npm run build-client`
 
-Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you’re on your own.
-
-You don’t have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn’t feel obligated to use this feature. However we understand that this tool wouldn’t be useful if you couldn’t customize it when you are ready for it.
-
-## Learn More
-
-You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
-
-To learn React, check out the [React documentation](https://reactjs.org/).
+Generates API client from the provided specification file into *src/generated/client*
