+++ package.json
@@ -3,36 +3,38 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "4.12.3",
-    "@material-ui/icons": "4.11.2",
-    "@reduxjs/toolkit": "1.6.1",
-    "@testing-library/jest-dom": "5.14.1",
-    "@testing-library/react": "11.2.7",
-    "@testing-library/user-event": "12.8.3",
-    "@types/jest": "26.0.24",
-    "@types/node": "14.17.12",
-    "@types/react": "17.0.19",
-    "@types/react-dom": "17.0.9",
-    "envalid": "7.2.1",
-    "immer": "9.0.5",
-    "keycloak-js": "15.0.2",
-    "localized-strings": "0.2.4",
-    "moment": "2.29.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-redux": "7.2.4",
-    "react-router-dom": "5.2.0",
+    "@emotion/react": "^11.4.1",
+    "@emotion/styled": "^11.3.0",
+    "@mui/icons-material": "^5.0.3",
+    "@mui/material": "^5.0.3",
+    "@reduxjs/toolkit": "^1.6.2",
+    "@testing-library/jest-dom": "^5.14.1",
+    "@testing-library/react": "^11.2.7",
+    "@testing-library/user-event": "^12.8.3",
+    "@types/jest": "^26.0.24",
+    "@types/node": "^12.20.28",
+    "@types/react": "^17.0.27",
+    "@types/react-dom": "^17.0.9",
+    "envalid": "^7.2.1",
+    "eslint-config-airbnb-typescript": "^14.0.1",
+    "keycloak-js": "^15.0.2",
+    "localized-strings": "^0.2.4",
+    "moment": "^2.29.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.5",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
-    "typescript": "4.4.2",
-    "web-vitals": "1.1.2"
+    "typescript": "^4.4.3",
+    "web-vitals": "^1.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "lint": "eslint --fix --ext .ts,.tsx ./src",
     "eject": "react-scripts eject",
-    "build-client": "openapi-generator-cli generate -i api-spec-template/swagger.yaml -o src/generated/client -c generator-config.json -g typescript-fetch"
+    "lint": "eslint --fix --ext .ts,.tsx ./src",
+    "build-client": "openapi-generator-cli generate -i api-spec/swagger.yaml -o src/generated/client -c generator-config.json -g typescript-fetch"
   },
   "eslintConfig": {
     "extends": [
@@ -53,15 +55,15 @@
     ]
   },
   "devDependencies": {
-    "@openapitools/openapi-generator-cli": "2.3.10",
-    "@types/react-router-dom": "5.1.8",
-    "@typescript-eslint/eslint-plugin": "4.29.3",
-    "@typescript-eslint/parser": "4.29.3",
-    "eslint": "7.32.0",
-    "eslint-config-airbnb-typescript": "12.3.1",
-    "eslint-plugin-import": "2.24.2",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "7.24.0",
-    "eslint-plugin-react-hooks": "4.2.0"
+    "@openapitools/openapi-generator-cli": "^2.4.12",
+    "@types/react-router-dom": "^5.3.1",
+    "@typescript-eslint/eslint-plugin": "^5.0.0",
+    "@typescript-eslint/parser": "^5.0.0",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb": "^18.2.1",
+    "eslint-plugin-import": "^2.25.1",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-react": "^7.26.1",
+    "eslint-plugin-react-hooks": "^4.2.0"
   }
 }
