+++ src/app/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import authReducer from "../features/auth/auth-slice";
+import authReducer from "../features/auth-slice";
 
 /**
  * Initialized Redux store
