+++ src/components/screens/main-screen.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import AppLayout from "components/layouts/app-layout";
-import { Typography } from "@material-ui/core";
+import { Typography } from "@mui/material";
 import strings from "localization/strings";
+import AppLayout from "components/layouts/app-layout";
 
 /**
  * Main screen component
