+++ src/features/auth-slice.ts
@@ -1,39 +1,33 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 import type { RootState } from "app/store";
 import { KeycloakInstance } from "keycloak-js";
-import { AccessToken } from "types";
 
 /**
- * Interface describing auth state in Redux
+ * Authentication state in Redux
  */
 export interface AuthState {
   keycloak?: KeycloakInstance;
-  accessToken?: AccessToken;
 }
 
 /**
- * Initial auth state
+ * Initial authentication state
  */
 const initialState: AuthState = {
-  keycloak: undefined,
-  accessToken: undefined
+  keycloak: undefined
 };
 
 /**
- * Auth slice of Redux store
+ * Authentication slice of Redux store
  */
 export const authSlice = createSlice({
   name: "auth",
   initialState: initialState,
   reducers: {
-    login: (state, { payload }: PayloadAction<AuthState>) => {
-      const { keycloak, accessToken } = payload;
-      state.keycloak = keycloak;
-      state.accessToken = accessToken;
+    login: (state, { payload }: PayloadAction<KeycloakInstance | undefined>) => {
+      state.keycloak = payload;
     },
     logout: state => {
       state.keycloak?.logout().then(() => {
-        state.accessToken = undefined;
         state.keycloak = undefined;
       });
     }
@@ -41,19 +35,19 @@ export const authSlice = createSlice({
 });
 
 /**
- * Auth actions from created auth slice
+ * Authentication actions from created authentication slice
  */
 export const { login, logout } = authSlice.actions;
 
 /**
- * Select Keycloak selector, used with useAppSelector custom hook defined for Redux store
+ * Select Keycloak selector
  *
  * @param state Redux store root state
  * @returns keycloak instance from Redux store
  */
-export const selectAuth = (state: RootState) => state.auth;
+export const selectKeycloak = (state: RootState) => state.auth.keycloak;
 
 /**
- * Reducer for auth slice
+ * Reducer from authentication slice
  */
 export default authSlice.reducer;
