+++ src/index.tsx
@@ -1,12 +1,15 @@
-import { CssBaseline, responsiveFontSizes, ThemeProvider } from "@material-ui/core";
-import { store } from "app/store";
 import React from "react";
 import ReactDOM from "react-dom";
-import { Provider } from "react-redux";
-import App from "components/App";
 import reportWebVitals from "reportWebVitals";
-import theme from "theme/theme";
+import { Provider } from "react-redux";
+import { store } from "app/store";
+import { CssBaseline, responsiveFontSizes, ThemeProvider } from "@mui/material";
+import theme from "theme";
+import App from "components/App";
 
+/**
+ * Initializes React component structure to the web page
+ */
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={ store }>
