+++ src/localization/fi.json
@@ -7,4 +7,5 @@
   "errorHandling": {
     "title": "Error"
   }
+
 }
