--- src/types/typings.d.ts
@@ -1,8 +1,0 @@
-declare module "@ckeditor/ckeditor5-react" {
-  export const CKEditor: any;
-}
-
-declare module "@ckeditor/ckeditor5-build-classic" {
-  const ClassicEditor: any;
-  export = ClassicEditor;
-}
